% Auto-generated: do not edit by hand
\name{dashEmojiMart}

\alias{dashEmojiMart}

\title{DashEmojiMart component}

\description{

}

\usage{
dashEmojiMart(id=NULL, autoFocus=NULL, categoryIcons=NULL, custom=NULL,
dynamicWidth=NULL, emojiButtonColors=NULL,
emojiButtonRadius=NULL, emojiButtonSize=NULL,
emojiSize=NULL, emojiVersion=NULL, exceptEmojis=NULL,
getSpritesheetURL=NULL, icons=NULL, locale=NULL,
maxFrequentRows=NULL, navPosition=NULL, noCountryFlags=NULL,
noResultsEmoji=NULL, onAddCustomEmoji=NULL,
onClickOutside=NULL, onEmojiSelect=NULL, perLine=NULL,
previewEmoji=NULL, previewPosition=NULL,
searchPosition=NULL, set=NULL, skin=NULL,
skinTonePosition=NULL, theme=NULL, value=NULL)
}

\arguments{
\item{id}{Character. }

\item{autoFocus}{Logical. }

\item{categoryIcons}{Named list. }

\item{custom}{Unnamed list. }

\item{dynamicWidth}{Logical. }

\item{emojiButtonColors}{Unnamed list. }

\item{emojiButtonRadius}{Character. }

\item{emojiButtonSize}{Numeric. }

\item{emojiSize}{Numeric. }

\item{emojiVersion}{Numeric. }

\item{exceptEmojis}{Unnamed list. }

\item{getSpritesheetURL}{}

\item{icons}{Character. }

\item{locale}{Character. }

\item{maxFrequentRows}{Numeric. }

\item{navPosition}{Character. }

\item{noCountryFlags}{Logical. }

\item{noResultsEmoji}{Character. }

\item{onAddCustomEmoji}{}

\item{onClickOutside}{}

\item{onEmojiSelect}{}

\item{perLine}{Numeric. }

\item{previewEmoji}{Character. }

\item{previewPosition}{Character. }

\item{searchPosition}{Character. }

\item{set}{Character. }

\item{skin}{Numeric. }

\item{skinTonePosition}{Character. }

\item{theme}{Character. }

\item{value}{Character. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

